# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mass Polimer ERP API",
    "version": "1.0.0",
...
    "license": {
      "name": "UNLICENSED"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This server"
    }
  ],
...
  "security": [
    {
      "FirebaseAuth": []
    },
    {
      "DevUser": []
    }
  ],
  "paths": {
...
    "/api/planning/orders": {
      "get": {
        "tags": [
          "Planning"
        ],
        "operationId": "listOrdersToPlan",
        "summary": "List orders awaiting planning",
        "description": "Confirmed orders still in `pending`, soonest delivery date first.\n\n**Requires permission:** `screen:orders_to_plan`",
        "parameters": [
          {
            "name": "x-dev-org",
            "in": "header",
            "required": false,
            "description": "Development only: pick the organization by id or slug when the identity has more than one membership.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The planning queue.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/SalesOrder"
                      },
                      {
                        "type": "object",
                        "properties": {
                          "customer": {
                            "$ref": "#/components/schemas/Customer",
                            "description": "A Customer; this endpoint may select only some of its fields."
                          }
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated` — no usable credential, or no membership resolved for it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — the caller's role does not hold `screen:orders_to_plan`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — unhandled server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/planning/operators": {
      "get": {
        "tags": [
          "Planning"
        ],
        "operationId": "listOperators",
        "summary": "List assignable operators",
        "description": "**Requires permission:** `screen:orders_to_plan`",
        "parameters": [
          {
            "name": "x-dev-org",
            "in": "header",
            "required": false,
            "description": "Development only: pick the organization by id or slug when the identity has more than one membership.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Active operators in the caller’s organization.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/Membership"
                      },
                      {
                        "type": "object",
                        "properties": {
                          "user": {
                            "$ref": "#/components/schemas/User",
                            "description": "A User; this endpoint may select only some of its fields."
                          }
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated` — no usable credential, or no membership resolved for it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — the caller's role does not hold `screen:orders_to_plan`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — unhandled server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans": {
      "get": {
        "tags": [
          "Planning"
        ],
        "operationId": "listPlans",
        "summary": "List production plans",
        "description": "**Requires permission:** `screen:plan_board`",
        "parameters": [
          {
            "name": "x-dev-org",
            "in": "header",
            "required": false,
            "description": "Development only: pick the organization by id or slug when the identity has more than one membership.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Production plans with machine, order and customer.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/ProductionPlan"
                      },
                      {
                        "type": "object",
                        "properties": {
                          "machine": {
                            "$ref": "#/components/schemas/Machine",
                            "description": "A Machine; this endpoint may select only some of its fields."
                          },
                          "salesOrder": {
                            "$ref": "#/components/schemas/SalesOrder",
                            "description": "A SalesOrder; this endpoint may select only some of its fields."
                          }
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated` — no usable credential, or no membership resolved for it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — the caller's role does not hold `screen:plan_board`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — unhandled server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Planning"
        ],
        "operationId": "createPlan",
        "summary": "Schedule an order onto a machine and shift",
        "description": "Moves the order to `planned`. One machine can hold only one plan per shift per day.\n\n**Requires permission:** `action:order.plan`",
        "parameters": [
          {
            "name": "x-dev-org",
            "in": "header",
            "required": false,
            "description": "Development only: pick the organization by id or slug when the identity has more than one membership.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "salesOrderId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "machineId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "shift": {
                    "type": "string",
                    "enum": [
                      "D",
                      "N"
                    ],
                    "default": "D"
                  },
                  "operatorName": {
                    "type": "string",
                    "default": ""
                  },
                  "scheduledStartDate": {
                    "type": "string",
                    "minLength": 1
                  },
                  "scheduledEndDate": {
                    "type": "string",
                    "default": ""
                  },
                  "logbookTemplateId": {
                    "type": "string"
                  }
                },
                "required": [
                  "salesOrderId",
                  "machineId",
                  "scheduledStartDate"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The scheduled plan.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ProductionPlan"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "machine": {
                          "$ref": "#/components/schemas/Machine",
                          "description": "A Machine; this endpoint may select only some of its fields."
                        },
                        "salesOrder": {
                          "$ref": "#/components/schemas/SalesOrder",
                          "description": "A SalesOrder; this endpoint may select only some of its fields."
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated` — no usable credential, or no membership resolved for it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — the caller's role does not hold `action:order.plan`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "- `not_found` — No order with that id exists in your organization.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "- `not_plannable` — The order is not awaiting planning.\n- `double_booked` — That machine, shift and day are already booked.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "- `bad_machine` — The referenced machine does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — unhandled server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/{id}/release": {
      "post": {
        "tags": [
          "Planning"
        ],
        "operationId": "releasePlan",
        "summary": "Release a plan back to the queue",
        "description": "Deletes the plan and returns its order to `pending`.\n\n**Requires permission:** `action:order.plan`",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Production plan id.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-dev-org",
            "in": "header",
            "required": false,
            "description": "Development only: pick the organization by id or slug when the identity has more than one membership.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Confirmation that the plan was released.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated` — no usable credential, or no membership resolved for it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — the caller's role does not hold `action:order.plan`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "- `not_found` — No plan with that id exists in your organization.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — unhandled server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "FirebaseAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase ID token, verified server-side. This is the production scheme; it activates when `DEV_AUTH=0` and the Firebase Admin credentials are configured."
      },
      "DevUser": {
        "type": "apiKey",
        "in": "header",
        "name": "x-dev-user",
        "description": "Development only, active while `DEV_AUTH` is on. Pass an employee code or email to act as that member; omit it to fall back to a seeded Administrator. Pair with `x-dev-org` to choose the organization. Never enable this in production."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "The envelope every failure is returned in.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Stable machine-readable code. Branch on this, not on the message."
              },
              "message": {
                "type": "string",
                "description": "Human-readable explanation, safe to surface in a UI."
              },
              "details": {
                "description": "Optional context. For `validation` this is the Zod field-error map."
              }
            }
          }
        }
      },
      "User": {
        "type": "object",
        "description": "The User record as stored. Relation objects appear only on endpoints that include them.",
        "properties": {
          "id": {
            "type": "string"
          },
          "firebaseUid": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
      "Membership": {
        "type": "object",
        "description": "The Membership record as stored. Relation objects appear only on endpoints that include them.",
        "properties": {
          "id": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "employeeCode": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "shift": {
            "type": "string"
          },
          "line": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "joinDate": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "lastSeen": {
            "type": "string"
          },
          "version": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "roleId": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "organizationId",
          "userId",
          "employeeCode",
          "department",
          "role",
          "shift",
          "line",
          "status",
          "joinDate",
          "location",
          "lastSeen",
          "version",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "Customer": {
        "type": "object",
        "description": "The Customer record as stored. Relation objects appear only on endpoints that include them.",
        "properties": {
          "id": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "gstNumber": {
            "type": "string"
          },
          "contactPerson": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "billingAddress": {
            "type": "string"
          },
          "deliveryAddress": {
            "type": "string"
          },
          "paymentTerms": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "version": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "organizationId",
          "name",
          "gstNumber",
          "contactPerson",
          "phone",
          "email",
          "billingAddress",
          "deliveryAddress",
          "paymentTerms",
          "status",
          "version",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "SalesOrder": {
        "type": "object",
        "description": "The SalesOrder record as stored. Relation objects appear only on endpoints that include them.",
        "properties": {
          "id": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "soNumber": {
            "type": "string"
          },
          "inquiryId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "product": {
            "type": "string"
          },
          "quantity": {
            "type": "integer"
          },
          "deliveryDate": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "specialInstructions": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "version": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "organizationId",
          "soNumber",
          "inquiryId",
          "customerId",
          "product",
          "quantity",
          "deliveryDate",
          "priority",
          "specialInstructions",
          "status",
          "version",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProductionPlan": {
        "type": "object",
        "description": "The ProductionPlan record as stored. Relation objects appear only on endpoints that include them.",
        "properties": {
          "id": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "salesOrderId": {
            "type": "string"
          },
          "machineId": {
            "type": "string"
          },
          "shift": {
            "type": "string"
          },
          "operatorName": {
            "type": "string"
          },
          "scheduledStartDate": {
            "type": "string"
          },
          "scheduledEndDate": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "logbookTemplateId": {
            "type": [
              "string",
              "null"
            ]
          },
          "version": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "organizationId",
          "salesOrderId",
          "machineId",
          "shift",
          "operatorName",
          "scheduledStartDate",
          "scheduledEndDate",
          "status",
          "version",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "Machine": {
        "type": "object",
        "description": "The Machine record as stored. Relation objects appear only on endpoints that include them.",
        "properties": {
          "id": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "line": {
            "type": "string"
          },
          "family": {
            "type": "string"
          },
          "logbookFormat": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "statusReason": {
            "type": [
              "string",
              "null"
            ]
          },
          "currentProduct": {
            "type": [
              "string",
              "null"
            ]
          },
          "currentFormula": {
            "type": [
              "string",
              "null"
            ]
          },
          "currentLot": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "organizationId",
          "code",
          "line",
          "family",
          "logbookFormat",
          "status"
        ]
      }
    }
  }
}